+++ scheduler/admin/job.py
@@ -36,7 +36,7 @@ class JobKwargInline(HiddenMixin, GenericStackedInline):
 
 
 class JobAdmin(admin.ModelAdmin):
-    actions = ['delete_model', 'disable_selected', 'enable_selected', 'run_job_now']
+    actions = ['delete_model', 'disable_selected', 'enable_selected', 'enqueue_job_now']
     inlines = [JobArgInline, JobKwargInline]
     list_filter = ('enabled',)
     list_display = ('enabled', 'name', 'job_id', 'function_string', 'is_scheduled',)
@@ -98,12 +98,11 @@ class JobAdmin(admin.ModelAdmin):
         level = messages.WARNING if not rows_updated else messages.INFO
         self.message_user(request, f"{message_bit} successfully enabled.", level=level)
 
-    @admin.action(description="Run now", permissions=('change',))
-    def run_job_now(self, request, queryset):
+    @admin.action(description="Enqueue now", permissions=('change',))
+    def enqueue_job_now(self, request, queryset):
         job_names = []
         for job in queryset:
-            job.schedule_now()
-            job.save()
+            job.enqueue_to_run()
             job_names.append(job.name)
         self.message_user(request, "The following jobs have been enqueued: %s" % (', '.join(job_names),))