+++ scheduler/apps.py
@@ -1,7 +1,7 @@
 from django.apps import AppConfig
 from django.utils.translation import gettext_lazy as _
 
-from scheduler.tools import reschedule_all_jobs, start_scheduler_thread
+from scheduler.tools import reschedule_all_jobs
 
 
 class SchedulerConfig(AppConfig):
@@ -16,4 +16,3 @@ class SchedulerConfig(AppConfig):
             # Django isn't ready yet, example a management command is being
             # executed
             pass
-        start_scheduler_thread()